+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Change Log
 
+## [0.13.2](https://github.com/bazel-contrib/vscode-bazel/compare/v0.13.1...v0.13.2) (2026-01-20)
+
+
+### Bug Fixes
+
+* **config:** Make default values for executables user-visible ([#489](https://github.com/bazel-contrib/vscode-bazel/issues/489)) ([529f5ad](https://github.com/bazel-contrib/vscode-bazel/commit/529f5ade95725b7f7b669e6832d473864c4d35f2))
+* **deps:** update dependency which to v6 ([#536](https://github.com/bazel-contrib/vscode-bazel/issues/536)) ([a7af391](https://github.com/bazel-contrib/vscode-bazel/commit/a7af3913ebaf6f2ded2a17c19cf824d31ea58621))
+
+
+### Performance Improvements
+
+* **workspace-tree:** implement debouncing and lazy loading for performance ([#542](https://github.com/bazel-contrib/vscode-bazel/issues/542)) ([8a534f8](https://github.com/bazel-contrib/vscode-bazel/commit/8a534f8d4e0c93cf1a4ab6f04c032dd7e0d82b8c)), closes [#504](https://github.com/bazel-contrib/vscode-bazel/issues/504)
+
 ## [0.13.1](https://github.com/bazel-contrib/vscode-bazel/compare/v0.13.0...v0.13.1) (2025-12-11)